Pulls the config commands into a separate cog. Also fixes some comments and other sylistic choices.
